Add papers on RAPTOR ATTACK
In this commit, I have uploaded crucial research papers and articles related to the "RAPTOR ATTACK" contribution in our group project. These papers delve into the specifics of the attack, including its methodology, potential vulnerabilities, and countermeasures. By including these papers in our repository, we aim to provide comprehensive resources for our team and the broader community interested in cybersecurity and deanonymization techniques.

The "RAPTOR ATTACK" is a fascinating yet menacing approach to deanonymizing Onion services, and these papers shed light on the techniques and challenges associated with this audacious attack. We encourage everyone to explore these resources to gain deeper insights into the realm of cybercrime investigation and the evolving landscape of Onion service security.

Let's continue working collaboratively to uncover the secrets of the Dark Web and strengthen our collective knowledge in the fascinating world of cybersecurity.

Happy reading and let's keep up the great work! 🎓🔍🌐